Refactored server.js to app.use to a router instead of programming all routes directly
