+++ server/middleware/mongoose.js
@@ -1,0 +1,15 @@
+const mongoose = require("mongoose");
+const connectDB = async () => {
+  try {
+    await mongoose.connect(process.env.MONGODB_URI, {
+      useUnifiedTopology: true,
+    });
+    console.log("MongoDB connected");
+  } catch (error) {
+    console.error("MongoDB connection error:", error);
+    process.exit(1);
+  }
+};
+connectDB();
+
+module.exports = mongoose;
