+++ server/middleware/s3.js
@@ -1,0 +1,10 @@
+const aws = require("aws-sdk");
+aws.config.update({
+  accessKeyId: process.env.AWS_ACCESS_KEY_ID,
+  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
+  region: process.env.AWS_REGION,
+});
+
+const s3 = new aws.S3();
+
+module.exports = s3;
